requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
sentence-transformers>=2.5.0
faiss-cpu>=1.7.4
pdfplumber>=0.10.0
camelot-py>=0.11.0
